Add warnings and error to actfw_raspberrypi.Display
